In the userpass section, the policies attached to a user are now listed (dev version)
Before it was not possible to see policies attached to userpass. 

For example in dev.go, in the part userpass, I attached a policy to fish1.
I had a line "policies" to dev.go where I put "default", "goldfish".

In the UI, I was able to see policies attached to my user (fish1) only in the login page. In the userpass section I do not see added policies.

To correct this I add code to userpass.go. The code permit to display policies attached to userpass.